[Color Picker] Keep alpha and fix the NaN hue in oklch output
The oklch string was built by hand from the L, C and H coordinates, which
dropped the alpha channel and wrote NaN as the hue for black, white and gray,
since those have no hue to report. ColorJS serializes both correctly, so the
formatter uses it instead.

Moved the formatter into its own module so the tests can reach it. utils.ts
imports @raycast/api, which ships types and no runtime, so nothing importing
it can be tested outside Raycast.
