+++ extensions/color-picker/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Fix] - {PR_MERGE_DATE}
 
 - Fix color swatches rendering as a grey square when the color format preference is set to a format that is not a valid CSS color, such as `HEX No Prefix`
+- Fix the `OKLCH` format dropping the alpha channel and producing a `NaN` hue for black, white and gray
 
 ## [Favorite Colors & AI Tools] - 2026-08-29
